refactor: add Doctor services and refactor commands
Introduce SeoDoctor and SiteDoctor domain services (src/Doctor/SeoDoctor.php, src/Doctor/SiteDoctor.php) to encapsulate site diagnosis and SEO audit logic. Refactor CLI commands (src/Command/Doctor.php, src/Command/DoctorSeo.php) to delegate to these services and simplify command code/path/status handling (e.g. formatStatus). Update library documentation (docs/8-Library.md and docs/8-Library.fr.md): refresh wording, add API link, add examples showing how to call the new Doctor services programmatically, and bump updated dates. This separates concerns so diagnostics can be reused outside the CLI and improves testability and maintainability.
